remove ts3.1 from node
