+++ types/node/ts3.4/globals.global.d.ts
