+++ types/node/ts3.4/node-tests.ts
@@ -1,31 +1,356 @@
-// tslint:disable-next-line:no-bad-reference
-import '../ts3.1/node-tests';
-import '../ts3.1/assert';
-import '../ts3.1/buffer';
-import '../ts3.1/child_process';
-import '../ts3.1/cluster';
-import '../ts3.1/crypto';
-import '../ts3.1/dgram';
-import '../ts3.1/global';
-import '../ts3.1/http';
-import '../ts3.1/http2';
-import '../ts3.1/net';
-import '../ts3.1/os';
-import '../ts3.1/path';
-import '../ts3.1/perf_hooks';
-import '../ts3.1/process';
-import '../ts3.1/querystring';
-import '../ts3.1/readline';
-import '../ts3.1/repl';
-import '../ts3.1/stream';
-import '../ts3.1/tls';
-import '../ts3.1/tty';
-import '../ts3.1/util';
-import '../ts3.1/worker_threads';
-import '../ts3.1/zlib';
+import * as fs from "fs";
+import * as url from "url";
+import * as util from "util";
+import * as http from "http";
+import * as https from "https";
+import * as net from "net";
+import * as console2 from "console";
+import * as timers from "timers";
+import * as inspector from "inspector";
+import * as trace_events from "trace_events";
 
-import { types, promisify } from 'util';
-import { BigIntStats, statSync, Stats } from 'fs';
+//////////////////////////////////////////////////////
+/// Https tests : http://nodejs.org/api/https.html ///
+//////////////////////////////////////////////////////
+
+{
+    let agent: https.Agent = new https.Agent({
+        keepAlive: true,
+        keepAliveMsecs: 10000,
+        maxSockets: Infinity,
+        maxFreeSockets: 256,
+        maxCachedSessions: 100,
+        timeout: 15000
+    });
+
+    agent = https.globalAgent;
+
+    let sockets: NodeJS.ReadOnlyDict<net.Socket[]> = agent.sockets;
+    sockets = agent.freeSockets;
+
+    https.request({
+        agent: false
+    });
+    https.request({
+        agent
+    });
+    https.request({
+        agent: undefined
+    });
+
+    https.get('http://www.example.com/xyz');
+    https.request('http://www.example.com/xyz');
+
+    https.get('http://www.example.com/xyz', (res: http.IncomingMessage): void => {});
+    https.request('http://www.example.com/xyz', (res: http.IncomingMessage): void => {});
+
+    https.get(new url.URL('http://www.example.com/xyz'));
+    https.request(new url.URL('http://www.example.com/xyz'));
+
+    https.get(new url.URL('http://www.example.com/xyz'), (res: http.IncomingMessage): void => {});
+    https.request(new url.URL('http://www.example.com/xyz'), (res: http.IncomingMessage): void => {});
+
+    const opts: https.RequestOptions = {
+        path: '/some/path'
+    };
+    https.get(new url.URL('http://www.example.com'), opts);
+    https.request(new url.URL('http://www.example.com'), opts);
+    https.get(new url.URL('http://www.example.com/xyz'), opts, (res: http.IncomingMessage): void => {});
+    https.request(new url.URL('http://www.example.com/xyz'), opts, (res: http.IncomingMessage): void => {});
+
+    https.globalAgent.options.ca = [];
+
+    {
+        function reqListener(req: http.IncomingMessage, res: http.ServerResponse): void {}
+
+        class MyIncomingMessage extends http.IncomingMessage {
+            foo: number;
+        }
+
+        class MyServerResponse extends http.ServerResponse {
+            foo: string;
+        }
+
+        let server: https.Server;
+
+        server = new https.Server();
+        server = new https.Server(reqListener);
+        server = new https.Server({ IncomingMessage: MyIncomingMessage});
+
+        server = new https.Server({
+            IncomingMessage: MyIncomingMessage,
+            ServerResponse: MyServerResponse
+        }, reqListener);
+
+        server = https.createServer();
+        server = https.createServer(reqListener);
+        server = https.createServer({ IncomingMessage: MyIncomingMessage });
+        server = https.createServer({ ServerResponse: MyServerResponse }, reqListener);
+
+        const timeout: number = server.timeout;
+        const listening: boolean = server.listening;
+        const keepAliveTimeout: number = server.keepAliveTimeout;
+        const maxHeadersCount: number | null = server.maxHeadersCount;
+        const headersTimeout: number = server.headersTimeout;
+        server.setTimeout().setTimeout(1000).setTimeout(() => {}).setTimeout(100, () => {});
+    }
+}
+
+/////////////////////////////////////////////////////
+/// Timers tests : https://nodejs.org/api/timers.html
+/////////////////////////////////////////////////////
+
+{
+    {
+        const immediate = timers
+            .setImmediate(() => {
+                console.log('immediate');
+            })
+            .unref()
+            .ref();
+        const b: boolean = immediate.hasRef();
+        timers.clearImmediate(immediate);
+    }
+    {
+        const timeout = timers
+            .setInterval(() => {
+                console.log('interval');
+            }, 20)
+            .unref()
+            .ref()
+            .refresh();
+        const b: boolean = timeout.hasRef();
+        timers.clearInterval(timeout);
+    }
+    {
+        const timeout = timers
+            .setTimeout(() => {
+                console.log('timeout');
+            }, 20)
+            .unref()
+            .ref()
+            .refresh();
+        const b: boolean = timeout.hasRef();
+        timers.clearTimeout(timeout);
+    }
+    async function testPromisify(doSomething: {
+        (foo: any, onSuccessCallback: (result: string) => void, onErrorCallback: (reason: any) => void): void;
+        [util.promisify.custom](foo: any): Promise<string>;
+    }) {
+        const setTimeout = util.promisify(timers.setTimeout);
+        let v: void = await setTimeout(100); // tslint:disable-line no-void-expression void-return
+        let s: string = await setTimeout(100, "");
+
+        const setImmediate = util.promisify(timers.setImmediate);
+        v = await setImmediate(); // tslint:disable-line no-void-expression
+        s = await setImmediate("");
+
+        // $ExpectType (foo: any) => Promise<string>
+        const doSomethingPromise = util.promisify(doSomething);
+
+        // $ExpectType string
+        s = await doSomethingPromise('foo');
+    }
+}
+
+/////////////////////////////////////////////////////////
+/// Errors Tests : https://nodejs.org/api/errors.html ///
+/////////////////////////////////////////////////////////
+
+{
+    {
+        Error.stackTraceLimit = Infinity;
+    }
+    {
+        const myObject = {};
+        Error.captureStackTrace(myObject);
+    }
+    {
+        const frames: NodeJS.CallSite[] = [];
+        Error.prepareStackTrace!(new Error(), frames);
+    }
+    {
+        const frame: NodeJS.CallSite = null!;
+        const frameThis: any = frame.getThis();
+        const typeName: string | null  = frame.getTypeName();
+        const func: Function | undefined  = frame.getFunction();
+        const funcName: string | null = frame.getFunctionName();
+        const meth: string | null  = frame.getMethodName();
+        const fname: string | null  = frame.getFileName();
+        const lineno: number | null  = frame.getLineNumber();
+        const colno: number | null  = frame.getColumnNumber();
+        const evalOrigin: string | undefined  = frame.getEvalOrigin();
+        const isTop: boolean = frame.isToplevel();
+        const isEval: boolean = frame.isEval();
+        const isNative: boolean = frame.isNative();
+        const isConstr: boolean = frame.isConstructor();
+    }
+}
+
+///////////////////////////////////////////////////////////
+/// Console Tests : https://nodejs.org/api/console.html ///
+///////////////////////////////////////////////////////////
+
+{
+    {
+        let _c: Console = console;
+        _c = console2;
+    }
+    {
+        const writeStream = fs.createWriteStream('./index.d.ts');
+        let consoleInstance: Console = new console.Console(writeStream);
+
+        consoleInstance = new console.Console(writeStream, writeStream);
+        consoleInstance = new console.Console(writeStream, writeStream, true);
+        consoleInstance = new console.Console({
+            stdout: writeStream,
+            stderr: writeStream,
+            colorMode: 'auto',
+            ignoreErrors: true
+        });
+        consoleInstance = new console.Console({
+            stdout: writeStream,
+            colorMode: false
+        });
+        consoleInstance = new console.Console({
+            stdout: writeStream
+        });
+    }
+    {
+        console.assert('value');
+        console.assert('value', 'message');
+        console.assert('value', 'message', 'foo', 'bar');
+        console.clear();
+        console.count();
+        console.count('label');
+        console.countReset();
+        console.countReset('label');
+        console.debug();
+        console.debug('message');
+        console.debug('message', 'foo', 'bar');
+        console.dir('obj');
+        console.dir('obj', { depth: 1 });
+        console.error();
+        console.error('message');
+        console.error('message', 'foo', 'bar');
+        console.group();
+        console.group('label');
+        console.group('label1', 'label2');
+        console.groupCollapsed();
+        console.groupEnd();
+        console.info();
+        console.info('message');
+        console.info('message', 'foo', 'bar');
+        console.log();
+        console.log('message');
+        console.log('message', 'foo', 'bar');
+        console.table({ foo: 'bar' });
+        console.table([{ foo: 'bar' }]);
+        console.table([{ foo: 'bar' }], ['foo']);
+        console.time();
+        console.time('label');
+        console.timeEnd();
+        console.timeEnd('label');
+        console.timeLog();
+        console.timeLog('label');
+        console.timeLog('label', 'foo', 'bar');
+        console.trace();
+        console.trace('message');
+        console.trace('message', 'foo', 'bar');
+        console.warn();
+        console.warn('message');
+        console.warn('message', 'foo', 'bar');
+
+        // --- Inspector mode only ---
+        console.profile();
+        console.profile('label');
+        console.profileEnd();
+        console.profileEnd('label');
+        console.timeStamp();
+        console.timeStamp('label');
+    }
+}
+
+/*****************************************************************************
+ *                                                                           *
+ * The following tests are the modules not mentioned in document but existed *
+ *                                                                           *
+ *****************************************************************************/
+
+///////////////////////////////////////////////////////////
+/// Inspector Tests                                     ///
+///////////////////////////////////////////////////////////
+
+{
+    {
+        const b: inspector.Console.ConsoleMessage = {source: 'test', text: 'test', level: 'error' };
+        inspector.open();
+        inspector.open(0);
+        inspector.open(0, 'localhost');
+        inspector.open(0, 'localhost', true);
+        inspector.close();
+        const inspectorUrl: string | undefined = inspector.url();
+
+        const session = new inspector.Session();
+        session.connect();
+        session.disconnect();
+
+        // Unknown post method
+        session.post('A.b', { key: 'value' }, (err, params) => {});
+        // TODO: parameters are implicitly 'any' and need type annotation
+        session.post('A.b', (err: Error | null, params?: {}) => {});
+        session.post('A.b');
+        // Known post method
+        const parameter: inspector.Runtime.EvaluateParameterType = { expression: '2 + 2' };
+        session.post('Runtime.evaluate', parameter,
+            (err: Error | null, params: inspector.Runtime.EvaluateReturnType) => {});
+        session.post('Runtime.evaluate', (err: Error, params: inspector.Runtime.EvaluateReturnType) => {
+            const exceptionDetails: inspector.Runtime.ExceptionDetails = params.exceptionDetails!;
+            const resultClassName: string = params.result.className!;
+        });
+        session.post('Runtime.evaluate');
+
+        // General event
+        session.on('inspectorNotification', message => {
+            message; // $ExpectType InspectorNotification<{}>
+        });
+        // Known events
+        session.on('Debugger.paused', (message: inspector.InspectorNotification<inspector.Debugger.PausedEventDataType>) => {
+            const method: string = message.method;
+            const pauseReason: string = message.params.reason;
+        });
+        session.on('Debugger.resumed', () => {});
+        // Node Inspector events
+        session.on('NodeTracing.dataCollected', (message: inspector.InspectorNotification<inspector.NodeTracing.DataCollectedEventDataType>) => {
+          const value: Array<{}> = message.params.value;
+        });
+    }
+}
+
+///////////////////////////////////////////////////////////
+/// Trace Events Tests                                  ///
+///////////////////////////////////////////////////////////
+
+{
+    const enabledCategories: string | undefined = trace_events.getEnabledCategories();
+    const tracing: trace_events.Tracing = trace_events.createTracing({ categories: ['node', 'v8'] });
+    const categories: string = tracing.categories;
+    const enabled: boolean = tracing.enabled;
+    tracing.enable();
+    tracing.disable();
+}
+
+////////////////////////////////////////////////////
+/// Node.js ESNEXT Support
+////////////////////////////////////////////////////
+
+{
+    const s = 'foo';
+    const s1: string = s.trimLeft();
+    const s2: string = s.trimRight();
+    const s3: string = s.trimStart();
+    const s4: string = s.trimEnd();
+}
 
 //////////////////////////////////////////////////////////
 /// Global Tests : https://nodejs.org/api/global.html  ///
@@ -39,10 +364,10 @@ import { BigIntStats, statSync, Stats } from 'fs';
 // Util Tests
 {
     const value: BigInt64Array | BigUint64Array | number = [] as any;
-    if (types.isBigInt64Array(value)) {
+    if (util.types.isBigInt64Array(value)) {
         // $ExpectType BigInt64Array
         const b = value;
-    } else if (types.isBigUint64Array(value)) {
+    } else if (util.types.isBigUint64Array(value)) {
         // $ExpectType BigUint64Array
         const b = value;
     } else {
@@ -50,15 +375,15 @@ import { BigIntStats, statSync, Stats } from 'fs';
         const b = value;
     }
 
-    const arg1UnknownError: (arg: string) => Promise<number> = promisify((arg: string, cb: (err: unknown, result: number) => void): void => { });
-    const arg1AnyError: (arg: string) => Promise<number> = promisify((arg: string, cb: (err: any, result: number) => void): void => { });
+    const arg1UnknownError: (arg: string) => Promise<number> = util.promisify((arg: string, cb: (err: unknown, result: number) => void): void => { });
+    const arg1AnyError: (arg: string) => Promise<number> = util.promisify((arg: string, cb: (err: any, result: number) => void): void => { });
 }
 
 // FS Tests
 {
-    const bigStats: BigIntStats = statSync('.', { bigint: true });
+    const bigStats: fs.BigIntStats = fs.statSync('.', { bigint: true });
     const bigIntStat: bigint = bigStats.atimeNs;
-    const anyStats: Stats | BigIntStats = statSync('.', { bigint: Math.random() > 0.5 });
+    const anyStats: fs.Stats | fs.BigIntStats = fs.statSync('.', { bigint: Math.random() > 0.5 });
 }
 
 // Global Tests
