+++ types/node/ts3.4/tslint.json
@@ -1,6 +1,10 @@
 {
-  "extends": "dtslint/dt.json",
-  "rules": {
-    "no-bad-reference": false
-  }
+    "extends": "dtslint/dt.json",
+    "rules": {
+        "ban-types": false,
+        "no-bad-reference": false,
+        "no-empty-interface": false,
+        "no-single-declare-module": false,
+        "unified-signatures": false
+    }
 }
