+++ types/node/ts3.6/index.d.ts
@@ -5,4 +5,4 @@
 /// <reference path="base.d.ts" />
 
 // tslint:disable-next-line:no-bad-reference
-/// <reference path="../ts3.1/assert.d.ts" />
+/// <reference path="../ts3.4/assert.d.ts" />
